update except urls

+++ code/run_server.sh
@@ -1,5 +1,16 @@
 #echo make sure you have python2.7 + django in your environment
+if [ -z $ESGF_NM_TASKS ]; then
+    echo please set ESGF_NM_TASKS to a directory for temporary task storage
+    exit 1
+fi
+
 rm $ESGF_NM_TASKS/*
+
+if [ -z $ESGF_NODEMGR_MAP ] ; then
+    echo please set ESGF_NODEMGR_MAP to your node map file
+    exit 1
+fi
+
 python server/basic_nodemanager.py $ESGF_NODEMGR_MAP $ESGF_NM_TIMESTAMP &
 server/nodemgr/manage.py runserver 0.0.0.0:80
 kill %1
